Clevo-P651RA: add notes from debugging
Spent a week trying to figure out why there is an infinite loop when the
card is turned back on using nouveau or bbswitch, but failed to figure
out a solution.

Workaround for now is to boot with acpi_osi="!Windows 2015".

This is weird, did the firmware author think there was a bug specific to
Windows 10? The faulty code is only executed for Windows 10 (identified
as _OSI("Windows 10")).

The DSDT and SSDT files can be found in dsl/Clevo_P651RA/.