Asus-X550ZE/atkd: add files and scripts for probing the ATK driver
Fn keys are not working for some reason, see
https://bugzilla.kernel.org/show_bug.cgi?id=115021

Under the suspicion that some WMI calls are missing, port the ATKD
device from the ASUS DSDT to a SSDT suitable that compiles and loads in
QEMU. Do not care if the behavior is incorrect, as long as stuff is
logged.

ppdebug.awk post-processes the asus.asl file to add lines to write to
the debug port. The LOGM methods were copied from QEMU (Seabios),
original name DBUG.

Run "make run" to test with Arch Linux (hard-coded path at the moment),
run "make run-win" to test with Windows 10 (arbitrary image which I
used in the past and was converted from an image downloaded from
https://developer.microsoft.com/en-us/microsoft-edge/tools/vms/ or
something).