d3test: use FW_CFG for VBIOS retrieval
The previous write to the XADR in the PCI Config space had no effect.
There were no vfio_pci_write_config traces visible. As a result, the
read data was also all zeroes and sometimes resulted in a new BSOD.
A _REG method showed that the config space was not ready at the time
_ROM was called so all reads/writes would fail.

Replace it by using FW_CFG to pass objects via IO ports, these are
always available. Now the device shows as successful in Device Manager
although Nvidia Display Settings fails with "You are not currently using
a display attached to an NVIDIA GPU" and the device does not enter D3.